Add support for blank lines
Alongside lined and dotted, this commit adds support for blank sections
in daily notes, reflect, and other note sections.

I wanted this because it's often convenient for me to zoom in a lot and
make really small notes in the daily notes section.

This configuration can be turned on by providing `blank: true` in
a configuration file (`dotted` must be false or not present).